refactor: Fusion des modules agriculture et production
- Suppression du module agriculture
- Enrichissement du module production avec les fonctionnalités agricoles
- Ajout des modèles pour cycles de culture et récoltes
- Mise à jour des schémas et endpoints
- Amélioration des statistiques de production
